+++ src/components/atoms/MarkdownBlock.jsx
@@ -3,11 +3,19 @@ import styled from 'styled-components'
 import { render_markdown } from '../../modules/markdown.js'
 
 const Block = styled.div`
+    width: 100%;
     max-width: 65ch;
+    min-width: 0;
     color: var(--ink);
+    overflow-wrap: anywhere;
 
     & > * + * {
         margin-top: 1rem;
+    }
+
+    pre {
+        max-width: 100%;
+        overflow-x: auto;
     }
 
     ul,